+++ Apps/CLI/Sources/PeekabooCLI/Commands/System/DialogCommand.swift
@@ -1,4 +1,3 @@
-import AXorcist
 import Commander
 import Foundation
 import PeekabooCore